requirements.txt
twitter-api-client
pandas